refactor: split EditorVote type into 2
